Delete the old version
