+++ src/index.js
@@ -1,10 +1,10 @@
-import updatedCssNamespace from './visitors/updatedCssNamespace';
+import cssNamespace from './visitors/cssNamespace';
 
 export default function() {
   return {
     visitor: {
       TaggedTemplateExpression(path, state) {
-        updatedCssNamespace(path, state);
+        cssNamespace(path, state);
       }
     }
   };
