+++ src/visitors/cssNamespace.js
@@ -1,122 +1,80 @@
 import * as t from 'babel-types';
-import { isStyled } from 'babel-plugin-styled-components/lib/utils/detectors';
+import {
+  isStyled,
+  isCSSHelper,
+  isInjectGlobalHelper,
+  isPureHelper,
+  isKeyframesHelper,
+  isHelper
+} from 'babel-plugin-styled-components/lib/utils/detectors';
+import { getCssNamespace } from './cssNamespace';
+import { loopWhile } from 'deasync';
+import postcss from 'postcss';
+import parentSelector from 'postcss-parent-selector';
+import unnest from 'postcss-nested';
 
-/**
- * Validates if a css rule contains a self reference "&". It is valid if
- * `.someClass & {` or `& {`, but, in case of `.myClass { ... } .myClass + & {`
- * it would be false as the regex requires the first rule be based on a self
- * reference.
- * @param {string} css
- */
-const startsWithSelfReference = css =>
-  /^(([^\{]([\w-_,:+\s]+))*\&\s*([^\{]([\w-_,:+\s]+))*)+[\{]/g.test(css.trim());
+const EXPRESSION = '_____';
 
-export const getCssNamespace = state => {
-  const { cssNamespace, rawCssNamespace } = state.opts;
-  if (!cssNamespace && !rawCssNamespace) {
-    return { cssNamespace: '&&', hasNamespaceSelfReference: true };
-  }
+const replacementNodes = new WeakSet();
 
-  const isCssNamespaceSet = !!cssNamespace;
-  const wrapperClass = isCssNamespaceSet
-    ? `.${[].concat(cssNamespace).join(' .')}`
-    : `${[].concat(rawCssNamespace).join(' &, ')}`;
-  const hasNamespaceSelfReference = startsWithSelfReference(
-    `${wrapperClass} {`
-  );
-  return { cssNamespace: wrapperClass, hasNamespaceSelfReference };
-};
-
-const wrapCssDefinition = (
-  cssNamespace,
-  css,
-  isSelfReferenced,
-  shouldLeaveScopeOpen
-) =>
-  isSelfReferenced
-    ? `${cssNamespace} {${css}${shouldLeaveScopeOpen ? '' : '}'}`
-    : `${cssNamespace} {&{${css}${shouldLeaveScopeOpen ? '' : '}}'}`;
-
-const NO_STRIP = false;
-const STRIP_END_CURLY_BRACKETS = true;
+// const customPlugin = postcss.plugin('customPlugin', config => )
 
 export default (path, state) => {
-  const { cssNamespace, hasNamespaceSelfReference } = getCssNamespace(state);
+  const { cssNamespace } = getCssNamespace(state);
 
-  if (
-    isStyled(path.node.tag, state) &&
-    path.node.quasi.quasis[0].value.cooked &&
-    !path.node.quasi.quasis[0].value.cooked.startsWith(`\n${cssNamespace} {`) &&
-    (path.node.tag.property
-      ? path.node.tag.property.name !== 'keyframes'
-      : true)
-  ) {
-    const { tag: callee, quasi: { quasis, expressions } } = path.node;
+  const { node } = path;
+  const { tag, quasi: { quasis, expressions } } = node;
 
-    let values;
-    if (quasis.length === 1) {
-      values = [
-        t.templateElement(
-          {
-            cooked: `\n${wrapCssDefinition(
-              cssNamespace,
-              quasis[0].value.cooked,
-              hasNamespaceSelfReference ||
-                startsWithSelfReference(quasis[0].value.cooked),
-              NO_STRIP
-            )}\n`,
-            raw: `\n${wrapCssDefinition(
-              cssNamespace,
-              quasis[0].value.raw,
-              hasNamespaceSelfReference ||
-                startsWithSelfReference(quasis[0].value.raw),
-              NO_STRIP
-            )}\n`
-          },
-          quasis[0].tail
-        )
-      ];
-    } else {
-      const isCookedSelfReferenced =
-        hasNamespaceSelfReference ||
-        startsWithSelfReference(quasis[0].value.cooked);
-      const isRawSelfReferenced =
-        hasNamespaceSelfReference ||
-        startsWithSelfReference(quasis[0].value.raw);
+  // Ignore nodes generated by this visitor, to prevent infinite loops
+  if (replacementNodes.has(node)) return;
 
-      const first = t.templateElement(
-        {
-          cooked: `\n${wrapCssDefinition(
-            cssNamespace,
-            quasis[0].value.cooked,
-            isCookedSelfReferenced,
-            STRIP_END_CURLY_BRACKETS
-          )}`,
-          raw: `\n${wrapCssDefinition(
-            cssNamespace,
-            quasis[0].value.raw,
-            isRawSelfReferenced,
-            STRIP_END_CURLY_BRACKETS
-          )}`
-        },
-        quasis[0].tail
-      );
+  // Ignore templates tagged with anything other than `styled(x)`
+  // Inspired by https://github.com/TrevorBurnham/babel-plugin-namespace-styled-components
+  if (!isStyled(t)(tag, state)) return;
+  if (isKeyframesHelper(t)(tag, state)) return;
+  if (isPureHelper(t)(tag, state)) return;
+  if (isHelper(t)(tag, state)) return;
+  if (isCSSHelper(t)(tag, state)) return;
+  if (isInjectGlobalHelper(t)(tag, state)) return;
+  if (tag.property && tag.property.name === 'keyframes') return; // Maintain backward compatibility with styled.keyframes
 
-      const last = t.templateElement(
-        {
-          cooked: `${quasis[quasis.length - 1].value
-            .cooked}${isCookedSelfReferenced ? '}' : '}}'}\n`,
-          raw: `${quasis[quasis.length - 1].value.raw}${isRawSelfReferenced
-            ? '}'
-            : '}}'}\n`
-        },
-        quasis[quasis.length - 1].tail
-      );
-      values = [first, ...quasis.slice(1, quasis.length - 1), last];
-    }
+  // Convert the tagged template to a string, with ${} expressions replaced with placeholders
+  const originalStyleString = quasis
+    .map(
+      (quasi, i) =>
+        expressions[i] ? quasi.value.raw + EXPRESSION : quasi.value.raw
+    )
+    .join('');
 
-    path.replaceWith(
-      t.taggedTemplateExpression(callee, t.templateLiteral(values, expressions))
-    );
+  const doesPrefixStartsWithSelfReference = cssNamespace.startsWith('&');
+  const prefix = doesPrefixStartsWithSelfReference ? cssNamespace : '&';
+
+  const processors = [unnest];
+  if (!doesPrefixStartsWithSelfReference) {
+    processors.push(parentSelector({ selector: cssNamespace }));
   }
+
+  let formattedCss = null;
+  postcss(processors)
+    .process(`\n${prefix} {${originalStyleString}}\n`, {
+      from: undefined // clears warning about SourceMap and Browserlist from postcss
+    })
+    .then(value => (formattedCss = value.css));
+  loopWhile(() => formattedCss === null);
+
+  const splitCss = formattedCss.split(EXPRESSION);
+  const values = splitCss.map((value, index) =>
+    t.templateElement(
+      { raw: value, cooked: value.replace(/\\\\/g, '\\') },
+      index === splitCss.length - 1
+    )
+  );
+
+  const replacementNode = t.taggedTemplateExpression(
+    tag,
+    t.templateLiteral(values, expressions)
+  );
+  replacementNodes.add(replacementNode);
+
+  path.replaceWith(replacementNode);
 };
